shoppingmall/requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.3
djangorestframework==3.9.4
drf-yasg==1.16.0
idna==2.8
importlib-metadata==0.18
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.1.0
packaging==19.0
pluggy==0.12.0
psycopg2==2.8.3
py==1.8.0
pyparsing==2.4.0
pytest==5.0.1
pytest-django==3.5.1
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
ruamel.yaml==0.15.98
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.5.2
